// src/app/Models/user.ts
export interface User {
    id_user: number;
    name: string;
    apellido: string;
    email: string;
    password: string;
    rol: string;
    barrio: string;
    direccion: string;
    edad: number;
    date_of_birth: string;
    document_type: string;
    document_number: number;
    city: string;
}